Converted all methods to static, commented authenticate
The original class was intended to be used as an object, but it makes more sense to use it as a class with static methods. The class does not need to store any data of its own, and the constructor's only other purpose was to version check openSSL.

The checking of openSSL would only ever be one off set up / debugging function and so for that reason I've moved this functionality to its own static method.

Next step is to write some example code and check for any bugs I may have introduced.